requirements.txt
flask>=2.0.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=0.19.0
numpy>=1.20.0
